S3 이미지 업로드 적용 및 기타 api 개발, 유저 인증 로직

+++ src/main/java/com/oss/gitborad/service/impl/CategoryServiceImpl.java
@@ -2,9 +2,11 @@ package com.oss.gitborad.service.impl;
 
 import com.oss.gitborad.data.domain.Category;
 import com.oss.gitborad.data.domain.CategoryGroup;
+import com.oss.gitborad.data.domain.User;
 import com.oss.gitborad.data.dto.CategoryDTO;
 import com.oss.gitborad.repository.CategoryGroupRepository;
 import com.oss.gitborad.repository.CategoryRepository;
+import com.oss.gitborad.repository.UserRepository;
 import com.oss.gitborad.service.CategoryService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -22,10 +24,12 @@ public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
     private final CategoryGroupRepository categoryGroupRepository;
+    private final UserRepository userRepository;
 
-    public CategoryServiceImpl(CategoryRepository categoryRepository, CategoryGroupRepository categoryGroupRepository) {
+    public CategoryServiceImpl(CategoryRepository categoryRepository, CategoryGroupRepository categoryGroupRepository, UserRepository userRepository) {
         this.categoryRepository = categoryRepository;
         this.categoryGroupRepository = categoryGroupRepository;
+        this.userRepository = userRepository;
     }
 
     //Category
@@ -45,6 +49,34 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
+    public List<CategoryDTO.Info> findListByWriter(Long id) {
+        User user = userRepository.getById(id);
+
+        Sort sort = Sort.by(Sort.Direction.DESC, "id");
+        Pageable pageable = PageRequest.of(0, categoryRepository.countByWriter(user), sort);
+        Page<CategoryDTO.Info> response = categoryRepository.findByWriter(user, pageable).map(x -> new CategoryDTO.Info(x));
+
+        List<CategoryDTO.Info> pageRequestDTO = new ArrayList<>();
+        for (CategoryDTO.Info i : response){
+            pageRequestDTO.add(i);
+        }
+        return pageRequestDTO;
+    }
+
+    @Override
+    public List<CategoryDTO.Info> findAll(int pageNumber, int size) {
+        Sort sort = Sort.by(Sort.Direction.DESC, "updateAT");
+        Pageable pageable = PageRequest.of(pageNumber, size , sort);
+        Page<CategoryDTO.Info> response = categoryRepository.findAll(pageable).map(x -> new CategoryDTO.Info(x));
+
+        List<CategoryDTO.Info> pageRequestDTO = new ArrayList<>();
+        for (CategoryDTO.Info i : response){
+            pageRequestDTO.add(i);
+        }
+        return pageRequestDTO;
+    }
+
+    @Override
     public void saveCategory(CategoryDTO.Request requestDTO) {
         Category category = Category.builder()
                 .name(requestDTO.getName())
@@ -55,7 +87,12 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void deleteCategory(Long id) {
+    public void deleteCategory(Long id, Long userId) {
+        Category category = categoryRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("카테고리를 찾을 수 없습니다."));
+        if(!category.getWriter().getId().equals(userId))
+            throw new RuntimeException("삭제 권한이 없습니다. user: "+userId);
+
         categoryRepository.deleteById(id);
     }
